# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Janiya API(End Point)",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendor/single/{id}/product": {
            "get": {
                "tags": [
                    "store"
                ],
                "summary": "check productby id",
                "description": "filter product by by id",
                "operationId": "vendorItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id of the item to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull pin request."
                    },
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad user Input"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/newAttribute": {
            "post": {
                "tags": [
                    "store"
                ],
                "summary": "Register new product",
                "description": "Register new record and return inserted data",
                "operationId": "newAttProduct",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "color",
                                    "size",
                                    "quantity"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "text"
                                    },
                                    "color": {
                                        "type": "integer"
                                    },
                                    "size": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "text"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad user Input"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}